threads: reformat sighandler and quitting
Readline, how about a very nice convenient function. I'm sure people would absolutely love it. It's called "readline_return_right_now_I_DEMAND_SO_I_AM_KING_YOU_BASTARD()". What it does is, it instantly unblocks readline and returns whatever, preferably NULL, but an incomplete line is fine too. Oh oh, and while we're at it, how bout another nice and convenient function called "readline_please_redraw_prompt_with_this(const char *)" or or "readline_auto_redraw(char **loc)".

+++ graph/graph_sdl.c
@@ -115,12 +115,12 @@ void graph_sdl_move_cam(graph_window *win)
 	if (win->flicked || win->translate) {
 		SDL_GetRelativeMouseState(&win->mousex, &win->mousey);
 		if (win->flicked) {
-			win->camera.view_roty += (float)win->mousex*MOUSE_ROT_SENS;
-			win->camera.view_rotx += (float)win->mousey*MOUSE_ROT_SENS;
+			win->camera.view_roty += win->mousex*MOUSE_ROT_SENS;
+			win->camera.view_rotx += win->mousey*MOUSE_ROT_SENS;
 		}
 		if (win->translate) {
-			win->camera.tr_x += -(float)win->mousex*MOUSE_TR_SENS;
-			win->camera.tr_y += (float)win->mousey*MOUSE_TR_SENS;
+			win->camera.tr_x += -win->mousex*MOUSE_TR_SENS;
+			win->camera.tr_y += win->mousey*MOUSE_TR_SENS;
 		}
 	}
 	if (win->chosen) {
